flaskSrc/requirements.txt
Flask==2.2.0
mysql-connector-python
gunicorn